+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+
+- brand typo
+
 ## [0.7.1] - 2021-10-18
 ### Fixed
 - add the /p parameter to the productList structured data URL